si puo tornare indietro alla challenge room

+++ Assets/Scripts/PlayerManager.cs
@@ -18,7 +18,7 @@ public class PlayerManager : NetworkBehaviour
     public int publicKeyEncode;
     [SyncVar]
     public int publicKeyModule;
-    public double password;
+    public int password;
     public int privateKey;
     public NetworkIdentity networkIdentity;
 
@@ -81,13 +81,13 @@ public class PlayerManager : NetworkBehaviour
         return id;
     }
 
-    public double getPassword()
+    public int getPassword()
     {
         return password;
     }
 
     [TargetRpc]
-    public void rcpTargetSetPassword(NetworkConnection target, double p)
+    public void rcpTargetSetPassword(NetworkConnection target, int p)
     {
         password = p;
     }
